add: Документация в navbar

+++ frontend/pages/api/worklist/[id]/index.ts
@@ -3,9 +3,11 @@ import { PrismaClient } from "@prisma/client";
 import { TaskCategoryDto } from "@/lib/dto/tasks";
 
 import { getDefaultHandler } from "@/lib/api/apiHandler";
-import { response } from "@/lib/api/response";
+import { response, responseAuth } from "@/lib/api/response";
 import { NextParsedUrlQuery } from "next/dist/server/request-meta";
 import { usePrisma } from "@/lib/api/database";
+import { ScopeEnum } from "@/lib/dto/users";
+import { WorklistDto } from "@/lib/dto/worklist";
 
 
 const prisma = usePrisma()
@@ -27,16 +29,68 @@ handler.get(response(async (req, res) => {
 })
 )
 
-
-handler.delete(response(async (req, res) => {
+handler.post(responseAuth(async (req, res, userId) => {
 	const id = (req.query as Query).id
 	if (!id) return { error: { code: 400, message: 'Неверный индекс' } }
 
-	await prisma.worklist.delete({
-		where: { id }
+	const user = await prisma.user.findUnique({
+		where: { id: userId },
+		include: { scopes: { include: { scope: true } } }
 	})
 
-	return { response: true }
+	if (user?.scopes.find(p => p.scope.value == ScopeEnum.admin)) {
+		const data = req.body as WorklistDto
+
+		const upset = await prisma.worklist.upsert({
+			where: { id },
+			update: {
+				name: data.name,
+				json: data.json,
+				date: new Date().toISOString()
+			},
+			create: {
+				name: data.name ? data.name : 'Безымянный',
+				json: data.json,
+				date: new Date().toISOString()
+			}
+		})
+
+		return { response: true }
+	} else {
+		return { error: { code: 403, message: "Данное действие доступно только администраторам" } }
+	}
+}))
+
+
+handler.delete(responseAuth(async (req, res, userId) => {
+	const user = await prisma.user.findUnique({
+		where: { id: userId },
+		include: { scopes: { include: { scope: true } } }
+	})
+
+	if (user?.scopes.find(p => p.scope.value == ScopeEnum.admin)) {
+		const id = (req.query as Query).id
+		if (!id) return { error: { code: 400, message: 'Неверный индекс' } }
+
+		await prisma.categoryToWorklist.deleteMany({
+			where: {
+				worklistId: id
+			}
+		})
+		await prisma.courseToWorklist.deleteMany({
+			where: {
+				worklistId: id
+			}
+		})
+
+		await prisma.worklist.delete({
+			where: { id }
+		})
+
+		return { response: true }
+	} else {
+		return { error: { code: 403, message: "Данное действие доступно только администраторам" } }
+	}
 
 }))
